implementing different kinds of tree branches to reduce file size

+++ FactoryTools/RegionVarCalculator_nr.h
@@ -17,13 +17,13 @@ private :
   //todo probably clean this up
   virtual EL::StatusCode doInitialize(EL::Worker * worker);
   virtual EL::StatusCode doCalculate (std::map<std::string, double>               & RegionVars,
-				      std::map<std::string, std::vector<double> > & VecRegionVars);
+				      std::map<std::string, std::vector<float> > & VecRegionVars);
 
   EL::StatusCode doAllCalculations   (std::map<std::string, double>               & RegionVars,
-				      std::map<std::string, std::vector<double> > & VecRegionVars);
+				      std::map<std::string, std::vector<float> > & VecRegionVars);
 
   EL::StatusCode doSRCalculations    (std::map<std::string, double>               & RegionVars,
-				      std::map<std::string, std::vector<double> > & VecRegionVars);
+				      std::map<std::string, std::vector<float> > & VecRegionVars);
 
 
 public :
