+++ Root/RegionVarCalculator_b4j.cxx
@@ -26,7 +26,7 @@ EL::StatusCode RegionVarCalculator_b4j::doInitialize(EL::Worker * worker) {
 }
 
 EL::StatusCode RegionVarCalculator_b4j::doCalculate(std::map<std::string, double              >& RegionVars,
-						     std::map<std::string, std::vector<double> >& VecRegionVars){
+						     std::map<std::string, std::vector<float> >& VecRegionVars){
   xAOD::TStore * store = m_worker->xaodStore();//grab the store from the worker
   xAOD::TEvent* event = m_worker->xaodEvent();
 
@@ -46,7 +46,7 @@ EL::StatusCode RegionVarCalculator_b4j::doCalculate(std::map<std::string, double
 
 
 EL::StatusCode RegionVarCalculator_b4j::doAllCalculations(std::map<std::string, double>& RegionVars,
-							   std::map<std::string, std::vector<double> > & VecRegionVars)
+							   std::map<std::string, std::vector<float> > & VecRegionVars)
 {
   xAOD::TStore * store = m_worker->xaodStore();
   xAOD::TEvent * event = m_worker->xaodEvent();
@@ -104,15 +104,15 @@ EL::StatusCode RegionVarCalculator_b4j::doAllCalculations(std::map<std::string, 
 
 
   //  const std::vector<xAOD::IParticle*> & jetStdVec = jetcont->stdcont();
-  std::vector<double> jetPtVec;
-  std::vector<double> jetEtaVec;
-  std::vector<double> jetPhiVec;
-  std::vector<double> jetEVec;
+  std::vector<float> jetPtVec;
+  std::vector<float> jetEtaVec;
+  std::vector<float> jetPhiVec;
+  std::vector<float> jetEVec;
 
-  std::vector<double> jetTau1Vec;
-  std::vector<double> jetTau2Vec;
-  std::vector<double> jetTau3Vec;
-  std::vector<double> jetDip12Vec;
+  std::vector<float> jetTau1Vec;
+  std::vector<float> jetTau2Vec;
+  std::vector<float> jetTau3Vec;
+  std::vector<float> jetDip12Vec;
 
   for( const auto& jet : *jets_nominal) {
     jetPtVec.push_back( toGeV(jet->pt()));
@@ -141,7 +141,7 @@ EL::StatusCode RegionVarCalculator_b4j::doAllCalculations(std::map<std::string, 
 
 
 EL::StatusCode RegionVarCalculator_b4j::doSRCalculations(std::map<std::string, double>& RegionVars,
-							  std::map<std::string, std::vector<double> > & VecRegionVars)
+							  std::map<std::string, std::vector<float> > & VecRegionVars)
 {
   return EL::StatusCode::SUCCESS;
 }
